moving chat application to a dedicated chat_app script
